# packages/backend/openapi.yaml
openapi: 3.1.0
info:
  title: Paints App API
  version: 1.0.0
  description: |
    Backend API for the **Paint Production & Inventory App** (spec:
    `doc/pr-1_spec.md`). This document is the single source of truth for
    request/response shapes — `openapi-typescript` generates
    `src/types/openapi.gen.ts` from it, and Scalar renders it at `/docs`.
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://service-dev.beautyflexinks.com
    description: Dev environment
  - url: https://service.beautyflexinks.com
    description: Production
...
security:
  - bearerAuth: []

# ============================================================================
# Components
# ============================================================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: ES256-signed JWT issued by `POST /auth/login`.
  parameters:
...
    Page:
      in: query
      name: page
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      in: query
      name: page_size
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      in: query
      name: search
      schema: { type: string, maxLength: 100 }
    IncludeArchived:
      in: query
      name: include_archived
      schema: { type: boolean, default: false }
  responses:
    Created:
      description: Resource created
      content:
        application/json:
          schema: { $ref: '#/components/schemas/IdRef' }
...
    Conflict:
      description: State machine violation or duplicate key
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    Classification:  { type: string, enum: [oil_based, water_based] }
    InkSeries:       { type: string, enum: [LCS, STD, OPQ_JS] }
    PoStatus:        { type: string, enum: [draft, ordered, shipped, received, cancelled] }
    PoLineKind:      { type: string, enum: [resource, finished_paint] }
...
    ProductionRunStatus:
      type: string
      enum: [planned, in_progress, completed, cancelled]
...
    Currency:          { type: string, pattern: '^[A-Z]{3}$', maxLength: 3, examples: [INR] }
...
    IdRef:
      type: object
      required: [id]
      properties: { id: { type: integer } }
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:   { type: string, examples: [Bad Request] }
        message: { type: string }
        code:    { type: string, description: 'Optional machine code (e.g. device_pending_approval)' }
        details: { type: array, items: { type: object } }
    Page:
      type: object
      required: [items, total, page, page_size]
      properties:
        items:     { type: array, items: { type: object } }
        total:     { type: integer }
        page:      { type: integer }
        page_size: { type: integer }

    # ---------- auth / users ----------
...
    PaintListItem:
      type: object
      required: [id, name]
      properties:
        id:              { type: integer }
        name:            { type: string }
        swatch:          { type: [string, 'null'] }
        notes:           { type: [string, 'null'] }
        hsn_code:        { type: [string, 'null'] }
        product_code:    { type: [string, 'null'] }
        tags:            { type: array, items: { type: string } }
        classifications: { type: array, items: { $ref: '#/components/schemas/Classification' } }
        ink_series:      { type: array, items: { $ref: '#/components/schemas/InkSeries' } }
        variant_count:   { type: [integer, string] }
        archived_at:     { type: [string, 'null'], format: date-time }
        created_at:      { type: string, format: date-time }
        updated_at:      { type: string, format: date-time }
...
    CreatePaintRequest:
      type: object
      required: [name, classifications, ink_series]
      properties:
        name:            { type: string, minLength: 1, maxLength: 255 }
        swatch:          { type: string, maxLength: 50 }
        notes:           { type: string }
        hsn_code:        { type: string, maxLength: 50 }
        product_code:    { type: string, maxLength: 50 }
        tags:            { type: array, items: { type: string, maxLength: 50 }, maxItems: 32 }
        classifications: { type: array, items: { $ref: '#/components/schemas/Classification' }, minItems: 1, maxItems: 2, uniqueItems: true }
        ink_series:      { type: array, items: { $ref: '#/components/schemas/InkSeries' },      minItems: 1, maxItems: 3, uniqueItems: true }
...
    FormulaListItem:
      type: object
      required: [id, variant_id, name, standard_output_kg]
      properties:
        id:                              { type: integer }
        variant_id:                      { type: integer }
        paint_id:                        { type: integer }
        paint_name:                      { type: string }
        classification:                  { $ref: '#/components/schemas/Classification' }
        ink_series:                      { $ref: '#/components/schemas/InkSeries' }
        name:                            { type: string }
        notes:                           { type: [string, 'null'] }
        standard_output_kg:              { type: [number, string] }
        is_default:                      { type: boolean }
        wastage_threshold_pct:           { type: [number, string, 'null'] }
        resource_variance_threshold_pct: { type: [number, string, 'null'] }
        dilution_threshold_pct:          { type: [number, string, 'null'] }
        archived_at:                     { type: [string, 'null'], format: date-time }
        created_at:                      { type: string, format: date-time }
        updated_at:                      { type: string, format: date-time }
...
    CreateFormulaRequest:
      type: object
      required: [variant_id, name, standard_output_kg, ingredients]
      properties:
        variant_id:                      { type: integer, minimum: 1 }
        name:                            { type: string, minLength: 1, maxLength: 255 }
        notes:                           { type: string }
        standard_output_kg:              { type: number, exclusiveMinimum: 0 }
        is_default:                      { type: boolean }
        wastage_threshold_pct:           { type: number, minimum: 0 }
        resource_variance_threshold_pct: { type: number, minimum: 0 }
        dilution_threshold_pct:          { type: number, minimum: 0 }
        ingredients:
          type: array
          minItems: 1
          maxItems: 200
          items:
            type: object
            required: [resource_id, quantity_kg]
            properties:
              resource_id: { type: integer, minimum: 1 }
              quantity_kg: { type: number, exclusiveMinimum: 0 }
...
    ResourceListItem:
      type: object
      required: [id, name]
      properties:
        id:                       { type: integer }
        name:                     { type: string }
        description:              { type: [string, 'null'] }
        aliases:                  { type: array, items: { type: string } }
        import_source:            { type: [string, 'null'] }
        current_stock_kg:         { type: [number, string] }
        weighted_avg_cost_per_kg: { type: [number, string] }
        low_stock_threshold_kg:   { type: [number, string, 'null'] }
        archived_at:              { type: [string, 'null'], format: date-time }
        created_at:               { type: string, format: date-time }
        updated_at:               { type: string, format: date-time }
...
    CreateResourceRequest:
      type: object
      required: [name]
      properties:
        name:                   { type: string, minLength: 1, maxLength: 255 }
        description:            { type: string }
        aliases:                { type: array, items: { type: string, maxLength: 100 }, maxItems: 32 }
        import_source:          { type: string, maxLength: 255 }
        low_stock_threshold_kg: { type: number, minimum: 0 }
...
    CustomerListItem:
      type: object
      required: [id, name]
      properties:
        id:               { type: integer }
        name:             { type: string }
        contact_name:     { type: [string, 'null'] }
        contact_phone:    { type: [string, 'null'] }
        contact_email:    { type: [string, 'null'] }
        gst_number:       { type: [string, 'null'] }
        default_currency: { $ref: '#/components/schemas/Currency' }
        archived_at:      { type: [string, 'null'], format: date-time }
        created_at:       { type: string, format: date-time }
...
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name:             { type: string, minLength: 1, maxLength: 255 }
        contact_name:     { type: string, maxLength: 255 }
        contact_phone:    { type: string, maxLength: 50 }
        contact_email:    { type: string, format: email, maxLength: 255 }
        billing_address:  { type: string }
        gst_number:       { type: string, maxLength: 20 }
        default_currency: { $ref: '#/components/schemas/Currency' }
        notes:            { type: string }
...
    SupplierPOC:
      type: object
      required: [name]
      properties:
        name:  { type: string, minLength: 1, maxLength: 255 }
        email: { type: string, format: email, maxLength: 255 }
        phone: { type: string, maxLength: 50 }
        role:  { type: string, maxLength: 100 }
    SupplierListItem:
      type: object
      required: [id, name]
      properties:
        id:           { type: integer }
        name:         { type: string }
        contact_name: { type: [string, 'null'] }
        email:        { type: [string, 'null'] }
        phone:        { type: [string, 'null'] }
        gst_number:   { type: [string, 'null'] }
        archived_at:  { type: [string, 'null'], format: date-time }
        created_at:   { type: string, format: date-time }
...
    CreateSupplierRequest:
      type: object
      required: [name]
      properties:
        name:         { type: string, minLength: 1, maxLength: 255 }
        contact_name: { type: string, maxLength: 255 }
        email:        { type: string, format: email, maxLength: 255 }
        phone:        { type: string, maxLength: 50 }
        address:      { type: string }
        website:      { type: string, maxLength: 255 }
        gst_number:   { type: string, maxLength: 20 }
        pocs:         { type: array, items: { $ref: '#/components/schemas/SupplierPOC' }, maxItems: 32 }
        notes:        { type: string }
...
    PurchaseOrderListItem:
      type: object
      required: [id, supplier_id, status, currency]
      properties:
        id:            { type: integer }
        supplier_id:   { type: integer }
        supplier_name: { type: string }
        status:        { $ref: '#/components/schemas/PoStatus' }
        currency:      { $ref: '#/components/schemas/Currency' }
        notes:         { type: [string, 'null'] }
        ordered_at:    { type: [string, 'null'], format: date-time }
        shipped_at:    { type: [string, 'null'], format: date-time }
        received_at:   { type: [string, 'null'], format: date-time }
        archived_at:   { type: [string, 'null'], format: date-time }
        created_at:    { type: string, format: date-time }
        total_cost:    { type: [number, string] }
        item_count:    { type: [integer, string] }
...
    POItemInput:
      type: object
      required: [kind, landed_cost_per_kg]
      properties:
        kind:               { $ref: '#/components/schemas/PoLineKind' }
        resource_id:        { type: integer, minimum: 1 }
        variant_id:         { type: integer, minimum: 1 }
        pack_size_kg:       { type: number, exclusiveMinimum: 0 }
        quantity_kg:        { type: number, exclusiveMinimum: 0 }
        quantity_packs:     { type: integer, exclusiveMinimum: 0 }
        landed_cost_per_kg: { type: number, minimum: 0 }
    CreatePORequest:
      type: object
      required: [supplier_id, items]
      properties:
        supplier_id: { type: integer, minimum: 1 }
        currency:    { $ref: '#/components/schemas/Currency' }
        notes:       { type: string }
        items:       { type: array, minItems: 1, maxItems: 200, items: { $ref: '#/components/schemas/POItemInput' } }
...
    ResourceActual:
      type: object
      required: [resource_id, expected_kg, actual_kg, flagged]
      properties:
        resource_id:   { type: integer }
        resource_name: { type: string }
        expected_kg:   { type: [number, string] }
        actual_kg:     { type: [number, string] }
        variance_pct:  { type: [number, string, 'null'] }
        flagged:       { type: boolean }
    DilutionAdjustment:
      type: object
      required: [id, resource_id, kg_added]
      properties:
        id:            { type: integer }
        resource_id:   { type: integer }
        resource_name: { type: string }
        kg_added:      { type: [number, string] }
        notes:         { type: [string, 'null'] }
        created_at:    { type: string, format: date-time }
...
    ResourceInventoryRow:
      type: object
      properties:
        id:                       { type: integer }
        name:                     { type: string }
        aliases:                  { type: array, items: { type: string } }
        current_stock_kg:         { type: [number, string] }
        weighted_avg_cost_per_kg: { type: [number, string] }
        low_stock_threshold_kg:   { type: [number, string, 'null'] }
        effective_threshold_kg:   { type: [number, string] }
        is_low_stock:             { type: boolean }

    # ---------- dashboard / reports ----------
...
    IrregularityReportRow:
      type: object
      properties:
        id:                  { type: integer }
        batch_number:        { type: string }
        variant_id:          { type: integer }
        formula_id:          { type: integer }
        status:              { $ref: '#/components/schemas/ProductionRunStatus' }
        expected_output_kg:  { type: [number, string] }
        actual_output_kg:    { type: [number, string, 'null'] }
        wastage_pct:         { type: [number, string, 'null'] }
        wastage_flagged:     { type: boolean }
        dilution_total_kg:   { type: [number, string] }
        dilution_flagged:    { type: boolean }
        operator:            { type: string }
        completed_at:        { type: [string, 'null'], format: date-time }
        resource_variances:  { type: array, items: { $ref: '#/components/schemas/ResourceActual' } }
        dilution_adjustments: { type: array, items: { $ref: '#/components/schemas/DilutionAdjustment' } }
...
paths:
...
  /paints:
    get:
      tags: [Paints]
      summary: List paints with pagination + filters
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeArchived'
        - { in: query, name: hsn_code, schema: { type: string } }
        - { in: query, name: product_code, schema: { type: string } }
        - { in: query, name: classification, schema: { $ref: '#/components/schemas/Classification' } }
        - { in: query, name: ink_series, schema: { $ref: '#/components/schemas/InkSeries' } }
        - { in: query, name: tag, schema: { type: string } }
      responses:
        '200':
          description: Page of paints
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/PaintListItem' } }
    post:
      tags: [Paints]
      summary: Create a paint + initial variant matrix (Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePaintRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /formulas:
    get:
      tags: [Formulas]
      summary: List formulas
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeArchived'
        - { in: query, name: variant_id, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Page of formulas
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/FormulaListItem' } }
    post:
      tags: [Formulas]
      summary: Create formula + ingredients (Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateFormulaRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
...
  /resources:
    get:
      tags: [Resources]
      summary: List resources
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeArchived'
        - { in: query, name: low_stock_only, schema: { type: boolean } }
      responses:
        '200':
          description: Page of resources
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/ResourceListItem' } }
    post:
      tags: [Resources]
      summary: Create a resource (Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateResourceRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
...
  /customers:
    get:
      tags: [Customers]
      summary: List customers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeArchived'
      responses:
        '200':
          description: Page of customers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/CustomerListItem' } }
    post:
      tags: [Customers]
      summary: Create (Manager or Sales)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCustomerRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
...
  /suppliers:
    get:
      tags: [Suppliers]
      summary: List suppliers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/IncludeArchived'
      responses:
        '200':
          description: Page of suppliers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/SupplierListItem' } }
    post:
      tags: [Suppliers]
      summary: Create
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSupplierRequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
...
  /purchase-orders:
    get:
      tags: [Purchase Orders]
      summary: List POs
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/IncludeArchived'
        - { in: query, name: status, schema: { $ref: '#/components/schemas/PoStatus' } }
        - { in: query, name: supplier_id, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Page of POs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/PurchaseOrderListItem' } }
    post:
      tags: [Purchase Orders]
      summary: Create a draft PO (Manager)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePORequest' }
      responses:
        '201': { $ref: '#/components/responses/Created' }
...
  /inventory/resources:
    get:
      tags: [Inventory]
      summary: Resource stock with low-stock flag
      parameters:
        - { in: query, name: low_stock_only, schema: { type: boolean } }
      responses:
        '200':
          description: Resources
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ResourceInventoryRow' } }

  # ---------- Dashboard ----------
...
  /api/losses:
    get:
      tags: [Reports]
      summary: Production Irregularity Report (Manager)
      parameters:
        - { in: query, name: from, schema: { type: string, format: date } }
        - { in: query, name: to, schema: { type: string, format: date } }
        - { in: query, name: variant_id, schema: { type: integer, minimum: 1 } }
        - { in: query, name: operator_id, schema: { type: integer, minimum: 1 } }
      responses:
        '200':
          description: Flagged runs
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/IrregularityReportRow' } }
...